Add base class, rename class to AsciiUtility
